experiments: add corpus_test2.py to generate audio of actual music
The music21 package includes a corpus of public domain works. This
script is my first attempt at interpreting the class structure of a work
in that corpus and making a reasonable facsimile of the audio.

There are still pops and clicks. I think this is due to each note being
a different individually generated waveform. This causes blips in the
edges of notes where the waveforms don't line up.