common.py: Convert the Defaults class into a singleton instance
There was a bug where if the framerate was changed, the generate pitches
would not correspond to the frequencies they should have been. Framerate
should not affect the pitch of the signal. It turned out that my
calculations were correct however, the Defaults class was not
implemented to behave like a singleton. As a result different modules
were using different values which doesn't work very well.